TASK-56737 Add Space Identifier when bookmarking a note (#554)
Prior to this change, the bookmarked note didn't had a space identifier attached into it which avoids bookmarks deletion when the space is deleted as well. This change introduces the space identifier in re-usable favorite button properties. In addition, a change has been made in WikiSpaceActivityPublisher to avoid duplicating favorite item for the activity.