feat: Extra margin to remove when CMS portlets are not displayed - MEED-2953-Meeds-io/MIPs#103 (#852)
Prior to this change, When we toggle between show/hide of the preview mode in the public page, an extra margin in displayed when the app is hidden.
This change allows to remove extra margin when app is hidden.

(cherry picked from commit 74cb82e10a036a51bedb63f1de370cdf07ecb017)