fix: Fix HTML codesnippet on edit - MEED-3013 - Meeds-io/meeds#1333 (#881)
Prior to this change, with the Meeds-io/MIPs#71 changes, the newly
introduced codesnippet plugin isn't well interpreted when editing the
RichEditor content. This change ensures to encode the content having '<'
and '>' characters correctly to make the plugin interpret the result
correctly.

At the same time, this will ensure to not save a draft version of the
content when editing only when modified and not all time when the first
time open it to edit.