FIX: Notes_basics actions_task: Move - Forbid to move a note below its children - EXO-74174-Meeds-io/meeds#2513
Prior to this fix, when a user moved a note that had children, he was
able to select the children and move the note below them. It creates an
infinite loop.
This commit filters the move treeview to not display children of the
current node, so the user will not be able to select any of them as a
destination.