Fix: fix unable to save snv translation - EXO-75758 - Meeds-io/meeds#2642 (#1238)
prior to this change, when editing an snv with advanced edition option
and adding a translation while automatic translation is disabled, the
save button remains disabled as it checks the validity of the note title
in the snv while it shouldn't in this case which causes the non
validation of the from to enable the save.
This PR ensures to ignore validating the title in an snv case to fix
thus issue.