fix: Fix refresh the whole page and just focus on note editor body when pressing enter on note editor title - EXO-73109 - Meeds-io/meeds#2742 (#1278)
Prior to this change when pressing enter on note editor title, the whole page is refreshed, the title is reseted and an error message is displayed when I write something. After this change, we ensure to not refresh anymore the whole page and that my cursor is focused on the note body.